Add import Wekan board feature
