+++ client/components/boards/boardHeader.jade
@@ -191,8 +191,14 @@ template(name="createBoard")
     input.primary.wide(type="submit" value="{{_ 'create'}}")
     span.quiet
       | {{_ 'or'}}
-      a(href="{{pathFor 'import'}}") {{_ 'import-board'}}
+      a.js-import-board {{_ 'import-board'}}
 
+template(name="chooseBoardSource")
+  ul
+    li
+      a(href="{{pathFor 'import/trello'}}") {{_ 'from-trello'}}
+    li
+      a(href="{{pathFor 'import/wekan'}}") {{_ 'from-wekan'}}
 
 template(name="boardChangeTitlePopup")
   form
