+++ i18n/en.i18n.json
@@ -145,6 +145,7 @@
     "computer": "Computer",
     "create": "Create",
     "createBoardPopup-title": "Create Board",
+    "chooseBoardSourcePopup-title": "Import board",
     "createLabelPopup-title": "Create Label",
     "current": "current",
     "date": "Date",
@@ -204,9 +205,13 @@
     "headerBarCreateBoardPopup-title": "Create Board",
     "home": "Home",
     "import": "Import",
-    "import-board": "import from Trello",
-    "import-board-title": "Import board from Trello",
-    "import-board-trello-instruction": "In your Trello board, go to 'Menu', then 'More', 'Print and Export', 'Export JSON', and copy the resulting text.",
+    "import-board": "import board",
+    "import-board-title-trello": "Import board from Trello",
+    "import-board-title-wekan": "Import board from Wekan",
+    "from-trello": "From Trello",
+    "from-wekan": "From Wekan",
+    "import-board-instruction-trello": "In your Trello board, go to 'Menu', then 'More', 'Print and Export', 'Export JSON', and copy the resulting text.",
+    "import-board-instruction-wekan": "In your Wekan board, go to 'Menu', then 'Export board', and copy the text in the downloaded file.",
     "import-json-placeholder": "Paste your valid JSON data here",
     "import-map-members": "Map members",
     "import-members-map": "Your imported board has some members. Please map the members you want to import to Wekan users",
